I commented out a line in resampleform.cpp that MSVC6 didn't like. I looked at it for a few minutes and I'm not sure why it didn't like it?? We'll come back to that when we implement resample form. I am checking in so that I can do some linux testing. I see a couple things that I think Linux will be mad about.

+++ mapimg/polyinv.c
@@ -1,4 +1,4 @@
-// $Id: polyinv.c,v 1.2 2005/01/27 18:56:21 jtrent Exp $
+// $Id: polyinv.c,v 1.3 2005/01/31 15:39:09 jtrent Exp $
 
 
 //Copyright 1993 United States Geological Survey
@@ -114,7 +114,7 @@ else
    b = al * al + (x/r_major) * (x/r_major);
    iflg = phi4z(es,e0,e1,e2,e3,al,b,&c,lat);
    if (iflg != OK)
-      return(iflg);
+      return (long)iflg;
    *lon = adjust_lon((asinz(x * c / r_major) / sin(*lat)) + lon_center);
    }
